requirements.txt
alembic
llama_index.core
llama_index.llms.ollama
llama_index.embeddings.ollama
llama_index.vector_stores.postgres
psycopg2-binary
openpyxl
fastapi
uvicorn[standard]